"Clearer" encrypted text plus a new message area
Because the Vigenere table allows the user to encrypt/decrypt the space
character, it is possible for the encrypted text to contain significant
white space. To help combat confusion, I added a beginning and ending
token to the encrypted text. This way, the encrypted message isn't
complete without the beginning and ending tokens and EVERYTHING in
between them.